refactor(ios): decompose IosPeripheral (326->240 lines) by extracting connect/disconnect/close/refreshServices
