+++ src/iosMain/kotlin/com/atruedev/kmpble/peripheral/IosPeripheralInternal.kt
@@ -6,7 +6,12 @@ import com.atruedev.kmpble.error.BleException
 import com.atruedev.kmpble.error.StaleGattHandle
 import com.atruedev.kmpble.gatt.Characteristic
 import com.atruedev.kmpble.gatt.Descriptor
+import com.atruedev.kmpble.gatt.DiscoveredService
 import com.atruedev.kmpble.gatt.internal.NotConnectedException
+import com.atruedev.kmpble.internal.StateRestorationHandler
+import com.atruedev.kmpble.peripheral.internal.PeripheralRegistry
+import kotlinx.coroutines.withContext
+import kotlinx.coroutines.withTimeout
 import platform.CoreBluetooth.CBCharacteristic
 import platform.CoreBluetooth.CBDescriptor
 import platform.Foundation.NSData
@@ -43,3 +48,53 @@ internal val L2CAP_OPEN_TIMEOUT = 30.seconds
 internal val DISCONNECT_TIMEOUT = 5.seconds
 internal val DISCOVERY_TIMEOUT = 10.seconds
 internal const val ATT_HEADER_SIZE = 3
+
+/**
+ * Lifecycle cleanup for [IosPeripheral].
+ *
+ * Extracted from the main class to keep it under 250 lines. Handles resource
+ * disposal, observer deregistration, and state restoration persistence.
+ */
+internal fun IosPeripheral.closeInternal() {
+    if (_closed.value) return
+    _closed.value = true
+    reconnectionHandler.stop()
+    bondManager.stop()
+    pairingRequestHandler.closeSync()
+    closeL2capChannels()
+    centralDelegate.unregisterConnectionCallback(identifier.value)
+
+    // Invalidate in-flight discovery cycle callbacks before teardown.
+    discoveryGeneration.incrementAndGet()
+    currentDiscovery = null
+    bridge.close()
+
+    observationManager.onObservationsChanged = null
+    observationManager.clear()
+    StateRestorationHandler.default.clearPersistedObservations(identifier.value)
+    peripheralContext.close()
+    PeripheralRegistry.remove(identifier)
+}
+
+/**
+ * Service discovery refresh for [IosPeripheral].
+ *
+ * Extracted from the main class to keep it under 250 lines.
+ */
+internal suspend fun IosPeripheral.refreshServicesInternal(): List<DiscoveredService> {
+    checkNotClosed()
+    return withContext(peripheralContext.dispatcher) {
+        val deferred = slots.armDiscovery()
+        // New discovery cycle: increment generation to invalidate stale callbacks
+        discoveryGeneration.incrementAndGet()
+        // Clear stale native handle mappings from previous cycle
+        nativeCharMap.clear()
+        nativeDescMap.clear()
+        bridge.discoverServices()
+        try {
+            withTimeout(currentTimeouts.serviceDiscovery) { deferred.await() }
+        } finally {
+            slots.clearDiscovery()
+        }
+    }
+}
